src/web_api/requirements.txt
azure-ai-documentintelligence==1.0.2
azure-core==1.33.0
azure-identity>=1.21.0
azure.monitor.opentelemetry>=1.6.9
azure-monitor-opentelemetry-exporter>=1.0.0b39
azure-search-documents>=11.5.2
azure-storage-blob==12.25.1
python-dotenv==1.1.1
fastapi==0.116.1
starlette==0.47.2
openai>=1.74.0
opentelemetry-api<=1.35.0
opentelemetry-exporter-otlp<=1.35.0
opentelemetry-instrumentation-httpx<=0.56b0
opentelemetry-instrumentation-logging<=0.56b0
opentelemetry-instrumentation-requests<=0.56b0
opentelemetry-sdk<=1.35.0
semantic-kernel==1.33.0
requests>=2.32.4
urllib3>=2.5.0
bs4==0.0.1
uvicorn==0.35.0
